+++ src/modapi/recipe/base/ICustomSmithingTrimRecipe.cpp
@@ -9,14 +9,15 @@ void ICustomSmithingTrimRecipe::_init() {
     auto tags = getCraftingTags();
     pImpl->mRecipes.reserve(tags.size());
     for (auto const& tag : tags) {
-        auto recipe = std::make_unique<::SmithingTrimRecipe>(
-            getRecipeId(),
-            getSmithingTemplate().pImpl->serialize(),
-            getBaseIngredient().pImpl->serialize(),
-            getAdditionIngredient().pImpl->serialize(),
-            HashedString(tag)
+        pImpl->mRecipes.push_back(
+            makeCustomSmithingRecipe<::SmithingTrimRecipe>(
+                tag,
+                getRecipeId(),
+                getSmithingTemplate(),
+                getBaseIngredient(),
+                getAdditionIngredient()
+            )
         );
-        pImpl->mRecipes.push_back(std::move(recipe));
     }
 }